# Swagger/swagg.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MovieTicketBooking(MTB) Application",
        "description": "API document for MovieTicketBooking Service",
        "x-task1-appname": "MTB",
        "x-task1-seal": 104437,
        "version": "23.03.0"
    },
...
    "servers": [
        {
            "url": "",
            "description": "http://localhost:1111-Relative server url",
            "variables": {}
        }
    ],
    "paths": {
...
        "/bookings/{bookingId}": {
            "get": {
                "tags": [
                    "BOOKINGS TABLE"
                ],
                "summary": "Get a booking by ID",
                "operationId": "getBookingById",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "BOOKINGS TABLE"
                ],
                "summary": "Update a booking",
                "operationId": "updateBooking",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Booking"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "BOOKINGS TABLE"
                ],
                "summary": "Delete a booking",
                "operationId": "deleteBooking",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking deleted successfully"
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/payments/{paymentId}": {
            "get": {
                "tags": [
                    "PAYMENTS TABLE"
                ],
                "summary": "Get a payment by ID",
                "operationId": "getPaymentById",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "PAYMENTS TABLE"
                ],
                "summary": "Update a payment",
                "operationId": "updatePayment",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Payment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "PAYMENTS TABLE"
                ],
                "summary": "Delete a payment",
                "operationId": "deletePayment",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment deleted successfully"
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Booking": {
                "type": "object",
                "properties": {
                    "BookingID": {
                        "type": "integer"
                    },
                    "NumberofSeats": {
                        "type": "integer"
                    },
                    "TimeStamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "Status": {
                        "type": "integer"
                    },
                    "Booking_UserID": {
                        "type": "string"
                    },
                    "Booking_ShowID": {
                        "type": "integer"
                    }
                }
            },
            "Payment": {
                "type": "object",
                "properties": {
                    "PaymentID": {
                        "type": "integer"
                    },
                    "AmountNumber": {
                        "type": "integer"
                    },
                    "TimeStamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "DiscountCoupon": {
                        "type": "string"
                    },
                    "RemoteTransactionID": {
                        "type": "string"
                    },
                    "PaymentMethod": {
                        "type": "string"
                    },
                    "PAYMENT_BookingID": {
                        "type": "integer"
                    }
                },
                "required": [
                    "PaymentID",
                    "AmountNumber",
                    "TimeStamp",
                    "PaymentMethod",
                    "PAYMENT_BookingID"
                ]
            },
...
            "ErrorInfo": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "string"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "bbt": {
                "type": "apiKey",
                "in": "header",
                "name": "BBT"
            },
            "token": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}